+++ csharp/tests/Valkey.Glide.IntegrationTests/StringCommandTests.cs
@@ -1,0 +1,83 @@
+// Copyright Valkey GLIDE Project Contributors - SPDX Identifier: Apache-2.0
+
+namespace Valkey.Glide.IntegrationTests;
+
+public class StringCommandTests(TestConfiguration config)
+{
+    public TestConfiguration Config { get; } = config;
+
+    internal static async Task GetAndSetValues(BaseClient client, string key, string value)
+    {
+        Assert.Equal("OK", await client.Set(key, value));
+        Assert.Equal(value, (await client.Get(key))!);
+    }
+
+    internal static async Task GetAndSetRandomValues(BaseClient client)
+    {
+        string key = Guid.NewGuid().ToString();
+        string value = Guid.NewGuid().ToString();
+        await GetAndSetValues(client, key, value);
+    }
+
+    [Theory(DisableDiscoveryEnumeration = true)]
+    [MemberData(nameof(Config.TestClients), MemberType = typeof(TestConfiguration))]
+    public async Task GetReturnsLastSet(BaseClient client) =>
+        await GetAndSetRandomValues(client);
+
+    [Theory(DisableDiscoveryEnumeration = true)]
+    [MemberData(nameof(Config.TestClients), MemberType = typeof(TestConfiguration))]
+    public async Task GetAndSetCanHandleNonASCIIUnicode(BaseClient client)
+    {
+        string key = Guid.NewGuid().ToString();
+        string value = "שלום hello 汉字";
+        await GetAndSetValues(client, key, value);
+    }
+
+    [Theory(DisableDiscoveryEnumeration = true)]
+    [MemberData(nameof(Config.TestClients), MemberType = typeof(TestConfiguration))]
+    public async Task GetReturnsNull(BaseClient client) =>
+        Assert.Null(await client.Get(Guid.NewGuid().ToString()));
+
+    [Theory(DisableDiscoveryEnumeration = true)]
+    [MemberData(nameof(Config.TestClients), MemberType = typeof(TestConfiguration))]
+    public async Task GetReturnsEmptyString(BaseClient client)
+    {
+        string key = Guid.NewGuid().ToString();
+        string value = string.Empty;
+        await GetAndSetValues(client, key, value);
+    }
+
+    [Theory(DisableDiscoveryEnumeration = true)]
+    [MemberData(nameof(Config.TestClients), MemberType = typeof(TestConfiguration))]
+    public async Task StrlenExistingKey(BaseClient client)
+    {
+        string key = Guid.NewGuid().ToString();
+        string value = Guid.NewGuid().ToString();
+        Assert.Equal("OK", await client.Set(key, value));
+
+        long result = await client.Strlen(key);
+        Assert.Equal(value.Length, result);
+    }
+
+    [Theory(DisableDiscoveryEnumeration = true)]
+    [MemberData(nameof(Config.TestClients), MemberType = typeof(TestConfiguration))]
+    public async Task StrlenNonExistingKey(BaseClient client)
+    {
+        string key = Guid.NewGuid().ToString();
+        long result = await client.Strlen(key);
+        Assert.Equal(0, result);
+    }
+
+    [Theory(DisableDiscoveryEnumeration = true)]
+    [MemberData(nameof(Config.TestClients), MemberType = typeof(TestConfiguration))]
+    public async Task StrlenWithUnicodeValue(BaseClient client)
+    {
+        string key = Guid.NewGuid().ToString();
+        string value = "שלום hello 汉字";
+        Assert.Equal("OK", await client.Set(key, value));
+
+        long result = await client.Strlen(key);
+        // Note: Strlen returns byte length, not character length
+        Assert.Equal(System.Text.Encoding.UTF8.GetByteCount(value), result);
+    }
+}
